extracted config_backup.go, config_restore.go
